refactor of NativeClass trait and init functions
Now all the needed information to register a class
to Godot is provided via trait implementations.

This is needed for implementing automation of binding-code generation
using procedural macros (or regular macros) properly without "mystifying"
the original written source code.

Another change along this is that the "owner" pointer is now no longer
only passed in the "constructor" in form of an opaque handle, but instead
the owning Godot-object has to be passed as a first parameter to every
function.