Add Instance API for creating NativeClass instances
When creating an game, it's often desired to return new nodes with custom Rust
behaviors attached. However, it's not previously possible to obtain a handle
to the current `GDNativeLibrary`, in order to create `NativeScript` instances,
through exposed APIs.

Fortunately, a pointer to the GDNativeLibrary being initialized is passed
through `godot_gdnative_init_options`. This commit saves this pointer in a
static, to use it later to instance scripted objects.

For more safety, the trait `Instanciable` is introduced as a generic bound.
`GDNativeLibrary::current_library` is added to the bindings as a type safe way
for users to get the object for their own uses.

The old, unused `NativeRef` type is replaced.