Manually drop arguments in generated bindings
Option<T> arguments in generated methods are previously dropped prematurely
due to move in pattern matching. This causes reference-counted objects to be
freed before the method is called, if the reference passed in is the only one.

Manual drops are added to prevent further bugs like this from compiling.

Fix #257

Co-authored-by: Max Clarke <maxeonyx@gmail.com>