Add workaround for iOS build
There is an issue in bindgen that prevents building for the aarch64-apple-ios target. This
works arounds the issue by replacing aarch64 with arm64 which clang can understand.

Also fixes conditional includes for macOS, which was using the target of the *build script*
instead of the target we're building for.

Close #285