Construct Godot API objects from class names
This adds the `Ref::<T>::by_class_name` constructor that instantiates a
sub-class of `T` by its name. If the resulting object cannot be casted to
`T`, it is correctly freed depending on whether it is a `Reference`, and
`None` is returned. As with `Ref::cast`, there is the caveat that reference-
counted classes cannot be casted to `Object`s, since memory management is
determined statically.