Move version detection out of user crate
This should make the `custom-godot` feature silence the warning
properly. Currently, the version requirement is hard-coded, but it
should be easy to later introduce a mechanism where the versions
requirements can be collected from `-sys` and `-bindings` and combined,
for better usefulness and less maintenance work in the future.

Close #1023